make code more idiomatic

+++ src/actions/setWindowTag.cpp
@@ -6,7 +6,7 @@ void setWindowTag(int tag) {
 		return;
 
 	XUnmapWindow(display, last_focused);
-	for (int i = 0; i < monitors[active_monitor].windows.size(); ++i) {
+	for (size_t i = 0; i < monitors[active_monitor].windows.size(); ++i) {
 		if (monitors[active_monitor].windows[i].win == last_focused) {
 			monitors[active_monitor].windows[i].tags = new_tag;
 			return;
