+++ src/onDestroyNotify.cpp
@@ -5,7 +5,7 @@
 
 void onDestroyNotify(const XDestroyWindowEvent &event) {
 #ifdef XINERAMA
-	for (int i = 0; i < amount_of_connected_monitors; ++i) {
+	for (size_t i = 0; i < amount_of_connected_monitors; ++i) {
 		for (auto ii = monitors[i].windows.begin(); ii != monitors[i].windows.end(); ++ii) {
 			if (ii->win == event.window) {
 				monitors[i].windows.erase(ii);
