+++ src/onMotionNotify.cpp
@@ -2,7 +2,7 @@
 
 void onMotionNotify(const XMotionEvent &event) {
 #ifdef XINERAMA
-	for (int i = 0; i < amount_of_connected_monitors; ++i) {
+	for (size_t i = 0; i < amount_of_connected_monitors; ++i) {
 		if (event.x >= monitors[i].x && event.x <= monitors[i].x + monitors[i].width && event.y >= monitors[i].y &&
 		    event.y <= monitors[i].y + monitors[i].height) {
 			active_monitor = i;
